feat: Add ClearValue::Color() default factory, fix parent setting and logging
- Added ClearValue::Color() static factory for default black (0,0,0,1) clear values
- Fixed prefab parent assignment: use Transform::parent directly instead of world.SetParent()
- Fixed scene serializer logging: cast fileSize to usize to avoid format warnings

+++ engine/scene/serialization/scene_serializer.cpp
@@ -205,7 +205,7 @@ bool SceneSerializer::LoadFromFile(const std::string& path, ecs::World& world) {
 
     bool result = LoadFromBuffer(buffer.data(), buffer.size(), world);
     if (result) {
-        NGE_LOG_INFO("Scene loaded from '{}' ({} bytes)", path, fileSize);
+        NGE_LOG_INFO("Scene loaded from '{}' ({} bytes)", path, static_cast<usize>(fileSize));
     }
     return result;
 }
